+++ client/Tests/Tests.jsx
@@ -52,6 +52,24 @@ const invalidArgDataType = props => {
   })`;
   return newTest;
 };
+const validMutation = (props) => {
+  const newTest = `describe('${props.testSuiteName}', () => {
+    it('${props.testDescription}', () => {
+    const validMutationTest = \`\n${props.writeTest}\`
+    tester.test(true, validMutationTest, ${props.writeInput})
+  })`
+  return newTest;
+}
+
+const invalidMutation = (props) => {
+  const newTest = `describe('${props.testSuiteName}', () => {
+    it('${props.testDescription}', () => {
+    const inValidMutationTest = \`\n${props.writeTest}\`
+    tester.test(false, validMutationTest, ${props.writeInput})
+  })`
+  return newTest;
+}
+
 
 export  {
     validQuery,
@@ -59,5 +77,7 @@ export  {
     validArgField,
     invalidArgField,
     validArgDataType,
-    invalidArgDataType
+    invalidArgDataType,
+    validMutation,
+    invalidMutation
 };
